Add "BrowserSelector" support
This adds a new mode that uses BrowserSelector as the host program,
instead of our host.go.

It's especially useful when we need a bidirectional link between IE and
Firefox. Since the URL pattern matching is done in BrowserSelector in
this mode, we no more need to maintain two separate copies of the URL
pattern definition.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>